Refactor; implement agda load on open/save
Unfortunately agda doesn't appear to support an editing model like
merlin where document damage is sent to the server incrementally. If the
file is loaded before save, it will be read from disk and return
incorrect highlight annotations. If we want to support highlighting
"on-the-fly" we will probably have to use temporary files.