Filter highlighting data for better stability
When a load command is issued to the agda interaction toplevel it
replies with several commands for clearing and annotating the editor
buffer.

Unfortunately, even when `NonInteractive` is specified, it does
not send the highlighting data in a single clean block. Instead, it
sends a large chunk up front, followed by several smaller incomplete
chunks, depending on whether or not the interface file is up to date.
Finally, it sends another big chunk that consists of all of the
previous data. The highlighting is supposed to be cleared at several
points during this process too.

I have no idea why agda does things this way since from the
editor point of view it's horribly inefficient. Highlighting annotations
would be applied synchronously in emacs but this is not the case for
vscode. When decorations are applied with vscode, it happens
asynchronously and is subject to several other buffer refreshing
actions. As a result, repeatedly resetting overlapping decorations in
rapid succession is likely to causing flickering or other glitches.

Indeed, before this patch, the highlighting would occassionally seem to
fail for no apparent reason but if the load command was triggered again
it might suddenly work again.

So what I have done is set up the points where highlighting is triggered
to execute the load command twice and simply ignore highlighting data
the first time around. This is an annoying hack but I don't see a
cleaner way to avoid firing several unnecessary decoration requests
without making the parser more stateful and ultimately more fragile.