Refactor highlights; now supports docs with holes
With the previous implementation of highlighting I was hoping that we
could avoid sending several decoration commands to the editor in rapid
succession since this isn't really appropriate for the asynchronous
approach that vscode uses.

To achieve this, the language server would load the file twice and
simply ignore highlighting commands from the first pass. The idea was
that the second pass would contain a single aggregate highlighting
command (in the response from agda) that we could execute.

However, it turns out that this is not always the case. In particular,
if the document contains holes or interaction points, the response from
agda still contains several separate highlighting commands even on the
second pass of the load command.

As a result, in order to get highlighting to work for documents that
have holes in them we need to actually deal with each of these separate
highlighting commands individually.

But the problem is that even a document of 50 lines or so might require
setting > 100 annotations. This problem is further compounded by the
fact that due to the way decorations are implemented, for efficiency
purposes we need to set them all at once for the document. So sending a
load command to agda for a document with around 50 lines with holes in
it might result in trying to set 25, 50, 75, and 100 annotations in
rapid succession (say, less than 10ms between requests to the editor).

So what I have done with this patch is set up the command to send
annotations to the editor to use a debouncing scheme. This more or less
does what I wanted the original implementation to do but hopefully in a
more robust way.

When the debounced annotation request is finally triggered and sent to
the editor it will contain the accumulation of all of the previous data
from the individual highlighting commands from agda.

There are two issues to investigate further. The first is what the
debounce threshold should be. But so far even just setting it to 0ms
seems to prevent multiple requests. The second issue is whether we
should (for efficiency purposes) try to dedupe the highlighting data
along the way. I'm not sure if agda will send redundant annotations or
not. I'm also not sure if it even matters if it does (i.e., the editor
might just ignore these anyway).